Add b&b code

+++ LPR381-Project/src/Models/Tester/Item.cs
@@ -8,15 +8,15 @@ namespace LPR381_Project.src.Models.Tester
 {
     internal class Item
     {
-        public float Weight { get; set; }
-        public float Value { get; set; }
+        public double Weight { get; set; }
+        public double Value { get; set; }
         public string Name { get; set; }
-        public float Ratio { get; set; }
+        public double Ratio { get; set; }
         public int Cofficient { get; set; }
         public int Chosen { get; set; }
-        public float Subtract { get; set; }
+        public double Subtract { get; set; }
 
-        public Item(string name, int coefficient, float value, float weight, int chosen, float subtract)
+        public Item(string name, int coefficient, double value, double weight, int chosen, double subtract)
         {
             this.Weight = weight;
             this.Value = value;
